General refactoring & minor bugfixing

+++ Prikhodko.NewsWebsite/Prikhodko.NewsWebsite.Service/CategoryService.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using System.Linq;
 using AutoMapper;
 using Prikhodko.NewsWebsite.Data.Contracts.Interfaces;
 using Prikhodko.NewsWebsite.Data.Contracts.Models;
@@ -19,35 +20,55 @@ namespace Prikhodko.NewsWebsite.Service
         }
         public void Add(CategoryServiceModel item)
         {
-            Category category = Mapper.Map<Category>(item); //TODO: add mapping profiles
+            if(item == null)
+            {
+                return;
+            }
+            Category category = Mapper.Map<Category>(item);
             repository.Add(category);
             unitOfWork.SaveChanges();
         }
 
         public void Delete(int id)
         {
+            if(id <= 0)
+            {
+                return;
+            }
             repository.Delete(id);
             unitOfWork.SaveChanges();
         }
 
         public CategoryServiceModel Get(int id)
         {
+            if (id <= 0)
+            {
+                return null;
+            }
+
             var category = repository.Get(id);
+
+            if(category == null)
+            {
+                return null;
+            }
+
             var result = Mapper.Map<CategoryServiceModel>(category);
             return result;
         }
 
         public IEnumerable<CategoryServiceModel> GetAll()
         {
-            var categories = repository.GetAll();
-            foreach (var category in categories)
-            {
-                yield return Mapper.Map<CategoryServiceModel>(category); //TODO: check whether this is correct
-            }
+            var categories = repository.GetAll().Select(x => Mapper.Map<CategoryServiceModel>(x));
+            return categories.ToList();
         }
 
         public void Update(CategoryServiceModel item)
         {
+            if(item == null)
+            {
+                return;
+            }
             var category = Mapper.Map<Category>(item);
             repository.Update(category);
             unitOfWork.SaveChanges();
